cn0540: Search gpios by labels
Instead of reading the gpio's attributes using the attribute number, a
labeled method was implemented. In this way, if the gpio order changes
in the devicetree, the plugin won't need additional changes.
Also the default gpio states requiered by the CN0540 are set at the end
of the init function.

Signed-off-by: Sergiu Cuciurean <sergiu.cuciurean@analog.com>